requirements.txt
markupsafe>=3.0.2
flask>=3.1.2
bcrypt>=4.3.0
flask-wtf>=1.2.2
python-dotenv>=1.1.1
flask-limiter>=3.13